Show more of the text field when editing.

+++ DoGood/DGFacebookManager.m
@@ -56,7 +56,7 @@
 }
 
 - (NSError *)responseErrorWithResponseCode:(NSInteger)code {
-    NSString *statusCodeMsg = [NSString stringWithFormat:@"The response status code is %ld", code];
+    NSString *statusCodeMsg = [NSString stringWithFormat:@"The response status code is %ld", (long)code];
     NSDictionary *errorInfo = @{
         NSLocalizedDescriptionKey: NSLocalizedString(@"Response error", nil),
         NSLocalizedFailureReasonErrorKey: NSLocalizedString(@"Access denied.", nil),
